edit profile of client prefill data

+++ lib/screens/client_screen/services/model/client_appointment_trans_model.dart
@@ -56,7 +56,7 @@ class Data {
   String? status;
   String? paymentNo;
   String? paymentMethod;
-  int? paymentAmount;
+  double? paymentAmount;
   String? paymentStatus;
   int? likhitDeCommission;
   double? gatewayCommission;
